pyproject.toml
[project]
name = "aiciv-mind"
version = "0.1.0"
requires-python = ">=3.12"
description = "Purpose-built AI operating system for AiCIV civilizations"
dependencies = [
    "anthropic>=0.86.0",
    "pyzmq>=26.0.0",
    "libtmux>=0.55,<0.56",
    "httpx>=0.27.0",
    "cryptography>=42.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "respx>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/aiciv_mind"]
